Feat: support <base> tags in websites
(note: these are quite rare)

Base tags can be specified, and when specified, they are expected to act as the new "base" URL for all other tags. We now support this by overwriting the old base URL with the newly defined one.

An example of a website using these base tags would be "https://dark.jakel.rocks" a rehost of the Netflix companion website for the show "Dark".