Fix: actually retry if there's errors :sweat_smile:
While looking into TLS errors not being retried and messing with NGINX, I realized that we just straight up weren't retrying errors on non-proxied connections. This fixes that, retrying any error up to the max limit. :tada: It also fixed the error message to be more in line withour others.