Handle default HQBatchSize in config
In Go, when a struct field of type int is not explicitly initialized, its default value is 0.
So, all config integers default to 0.
Then, in HQ producer we have a special function to set its default value
to 100 if its 0.
We can simplify this using viper. We set the default HQBatchSize to 100
in ConfigInit and drop the redundant function.

We also add a small unit test to validate config init.