Move closeBody(seed *model.Item) to Item.Close() (#334)
`closeBody` is a function about the internal state of the `Item`.
It makes sense to put it n `pkg.models.item` and rename it as `Close`.
Its better to have all the `Item` related methods in one place in
`pkg.models.Item` instead of having them spread in preprocessor /
postprocessor code.