Make config more unit-test friendly
The Config struct is stored in a local var and returned via `config.Get()`.
If you need to make a unit test that includes e.g. `config.Get().DisableHTMLTag`
it is difficult to mock the config object. This is not friendly to
unit-testing.

We rename `config` to `AppConfig` in `internal/pkg/config/config.go`.
This way, its possible to use `AppConfig.DisableHTMLTag` in other
modules and its easy to mock it in unit tests.

To showcase the new syntax, we change the HTML extractor to use the new
`AppConfig`.

The old syntax isn't affected.