Improve URL mimetype / content-type handling (#344)
* Define URL.mimetype from Content-Type HTTP header

If URL.mimetype is not set, check HTTP response `Content-Type` header
and try to define it from that.

Add unit tests to showcase the robustness of mimetype detection.

* Stop using Content-Type HTTP header and use URL.mimetype instead

Now that we define URL.mimetype from the `Content-Type`, there is no
need to lookup the HTTP header in every extractor.

* Use mime.ParseMediaType in URL.GetMIMEType

* First check HTTP Content-Type and then fallback to mimetype

Change the priority for mimetype detection.

* Add Content-Type to tests

It is necessary for mock HTTP responses to have HTTP header
`Content-Type` now.
This isn't bad, all real HTTP responses have this HTTP header.

* Fix css extractor IsCSS error

* Fix typo