Move SetDocument(nil) to Item.Close method
Placing `SetDocument(nil)` in the end of the method doesn't guarantee
that it will be executed always. If the method returns early, it is
skipped and memory isn't freed.

`Item` has a `Close()` method. Conceptually its better to include memory
cleanup in `Item.Close`.

Also, method `postprocessItem` includes `defer item.Close()`.

So, it makes sense to move `SetDocument(nil)` in there.