Return error instead of using panic() in the Reactor (#475)
Methods `ReceiveFeedback` and `ReceiveInsert` already return `error`.

We call `reactor.ReceiveFeedback` in the finisher and we terminate when
there is an error so the program behavior doesn't change.
https://github.com/internetarchive/Zeno/blob/main/internal/pkg/finisher/finisher.go#L122

We call `reactor.ReceiveInsert` in the pipeline and we terminate when
there is an error so the program behavior doesn't change.
https://github.com/internetarchive/Zeno/blob/main/internal/pkg/controler/pipeline.go#L161

`models.ErroNotASeed` is already defined, there is no need to make a new
error.

Define `ErrReactorItemPresent` because there is no relevant reactor
error.

Relevant to issue: https://github.com/internetarchive/Zeno/issues/473