Optimise internal/pkg/postprocessor/extractor/html_document_test.go (#485)
Two optimisations:
We iterate query `key` -> `values` and then iterate each `value` in `values`.
We encoded the `key` again and again for each `value`. We only need to
do it once. We move the key encoding code outside the `values` loop.

We define `enc.NewEncoder().String()` every time we need to encode a
value. We can define `encoder := enc.NewEncoder()` in the beginning and
reuse it.

We add a unit test for a bad URL just to be sure.